refactor: AdminStoreHydration 파일명을 AdminAuthGuard로 변경(#460)
